created auth lib file which hosts auth route logic and updated middleware and server side fetch parcels file to append auth token to header
