+++ src/lib/fetch/fetchParcels.ts
@@ -1,22 +1,28 @@
-import { ENDPOINTS, CONTEXTS } from '@/lib/constants';
-import { AgriParcel } from '@/lib/interfaces';
+import { ENDPOINTS, CONTEXTS } from "@/lib/constants";
+import { AgriParcel } from "@/lib/interfaces";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/lib/auth";
 
-export const dynamic = 'force-dynamic'
+export const dynamic = "force-dynamic";
 
 export async function fetchParcels(): Promise<AgriParcel[]> {
+  const session = await getServerSession(authOptions);
+  const token = session?.user?.accessToken;
+
   const searchParams = {
-    type: 'AgriParcel',
-    limit: '10',
+    type: "AgriParcel",
+    limit: "10",
   };
 
   const urlQuery = new URLSearchParams(searchParams).toString();
-  const url = new URL(`${ENDPOINTS.API_BASE_URL}?${urlQuery}`).toString();
+  const url = `${ENDPOINTS.API_BASE_URL}?${urlQuery}`;
 
   const response = await fetch(url, {
     headers: {
       Link: CONTEXTS.AGRIFARM,
+      Authorization: token ? `Bearer ${token}` : "",
     },
-    cache: 'no-store'
+    cache: "no-store",
   });
 
   if (!response.ok) {
